update readmes and full pipeline for meetit

+++ dataset_generation/pre_generation/download_librispeech.sh
@@ -7,10 +7,15 @@ declare -a sets=(test-clean.tar.gz train-clean-100.tar.gz train-clean-360.tar.gz
 
 for set in ${sets[@]}
 do
-    echo "Downloading "${set}
-    wget http://www.openslr.org/resources/12/${set} -P ${data_dir}
-    echo "Extracting "${set}
-    tar -xf ${data_dir}${set} -C ${data_dir}
-    rm ${data_dir}${set}
+	dir_name=$(basename ${set} .tar.gz)
+	if [[ -d "${data_dir}${dir_name}" ]]; then	# Check directory has not already been downloaded
+		echo "${data_dir}${dir_name} was already downloaded"
+	else
+		echo "Downloading "${set}
+		wget http://www.openslr.org/resources/12/${set} -P ${data_dir}
+		echo "Extracting "${set}
+		tar -xf ${data_dir}${set} -C ${data_dir}
+		rm ${data_dir}${set}
+	fi
 done
