+++ dataset_generation/pre_generation/download_noises_from_zenodo.sh
@@ -4,7 +4,11 @@
 data_dir=../../dataset/freesound/
 mkdir -p ${data_dir}
 
-wget https://zenodo.org/record/4019030/files/disco_noises.zip -P ${data_dir}
-unzip -q ${data_dir}disco_noises.zip -d ${data_dir}
-mv ${data_dir}disco_noises/ ${data_dir}data/
-rm ${data_dir}disco_noises.zip
+if [[ -d "${data_dir}data/" ]]; then
+	echo "data has already been downloaded"
+else
+    wget https://zenodo.org/record/4019030/files/disco_noises.zip -P ${data_dir}
+    unzip -q ${data_dir}disco_noises.zip -d ${data_dir}
+    mv ${data_dir}disco_noises/ ${data_dir}data/
+    rm ${data_dir}disco_noises.zip
+fi
